Rename vestigial DBUser instances to SourceUser
